refactor: split detector benchmark command modules
Keep dam-detect-bench main wiring small by moving CLI parsing and benchmark evaluation into focused modules with sibling test files.
